requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
groq==0.4.1
python-dotenv==1.0.0